move archivereference properties to correct folder
